refactor out the unmute loop into the mute manager
